AutoOnline/requirements.txt
cffi==1.6.0
cryptography==1.3.1
Django==1.9.5
docker-py==1.8.0
ecdsa==0.13
enum34==1.1.4
Fabric==1.11.1
idna==2.1
ipaddress==1.0.16
paramiko==1.17.0
py2-ipaddress==3.4.1
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
PyMySQL==0.7.2
requests==2.10.0
six==1.10.0
websocket-client==0.37.0
